LOG&KPI/requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.20.0
pandas>=1.3.0
scipy>=1.7.0
tqdm>=4.60.0
pyyaml>=6.0

# Optional: for MCP server
mcp>=0.1.0

